Refactored the code to make it more maintainable.
- Redid the code for the cloud to be able to specify subnets, with the default based on bridge number. It now adds a bridge every time you run it.

- The creating if the VM now prompts you for a bridge to select if there is more than one, and prompts you to select storage if there is more than 1

- The No Subcriber nag removal will now last through Proxmox updates

+++ tools/CREATE-ISO-macOS.command
@@ -7,6 +7,8 @@
 # https://luchina.com.br
 # 20211116T2245
 
+set -e
+
 clear
 
 echo -e "\nAutomate script for create \"ISO\" file of macOS Install in Proxmox VE Environament"
@@ -14,21 +16,27 @@ echo -e "BY: https://luchina.com.br"
 echo -e "SUPPORT: https://osx-proxmox.com"
 
 echo -n -e "\nPath to temporary files (work dir): "
-read TEMPDIR
+read -r TEMPDIR
 
 echo -n -e "Path to macOS Installation (.app) file: "
-read APPOSX
+read -r APPOSX
 
 echo " "
 
 ## Core 
-cd ${TEMPDIR} > /dev/null 2> /dev/null
-rm -rf macOS-install* > /dev/null 2> /dev/null
-hdiutil create -o macOS-install -size 16g -layout GPTSPUD -fs HFS+J > /dev/null 2> /dev/null
-hdiutil attach -noverify -mountpoint /Volumes/install_build macOS-install.dmg > /dev/null 2> /dev/null
-sudo "${APPOSX}/Contents/Resources/createinstallmedia" --volume /Volumes/install_build --nointeraction
-hdiutil detach -force "/Volumes/Install macOS"* > /dev/null 2> /dev/null && sleep 3s > /dev/null 2> /dev/null
-hdiutil detach -force "/Volumes/Shared Support"* > /dev/null 2> /dev/null
-mv macOS-install.dmg macOS-install.iso > /dev/null 2> /dev/null
+if [ -d "$TEMPDIR" ]; then
+    cd "$TEMPDIR" || exit 1
+    rm -rf macOS-install* > /dev/null 2> /dev/null
+    hdiutil create -o macOS-install -size 16g -layout GPTSPUD -fs HFS+J > /dev/null 2> /dev/null
+    hdiutil attach -noverify -mountpoint /Volumes/install_build macOS-install.dmg > /dev/null 2> /dev/null
+    sudo "${APPOSX}/Contents/Resources/createinstallmedia" --volume /Volumes/install_build --nointeraction
+    hdiutil detach -force "/Volumes/Install macOS"* > /dev/null 2> /dev/null && sleep 3s > /dev/null 2> /dev/null
+    hdiutil detach -force "/Volumes/Shared Support"* > /dev/null 2> /dev/null
+    mv macOS-install.dmg macOS-install.iso > /dev/null 2> /dev/null
+else
+    echo "The temporary directory does not exist!"
+    exit 1
+fi
 
 echo " "
+
